+++ packages/sveltekit-sdk/src/lib/server/index.ts
@@ -13,4 +13,4 @@
 import '$app/server'
 export { createMedusaHandle } from './hooks'
 export { requestContext as getMedusaContext } from './request'
-export { forwardAnalytics, analyticsPOST, setTraits } from './analytics'
+export { forwardAnalytics, analyticsPOST, setTraits } from './events'