+++ packages/js-sdk/CHANGELOG.md
@@ -1,12 +1,16 @@
-## 2.17.2e
+## 2.17.2-f
+
+- Rename createAnalyticsCollector to createCollector
+
+## 2.17.2-e
 
 - Add coverage for medusa-plugin-reviews, medusa-plugin-content, medusa-plugin-affiliates, and medusa-plugin-analytics
 
-## 2.17.2c
+## 2.17.2-c
 
 - Fix type error in forms route
 
-## 2.17.2b
+## 2.17.2-b
 
 - Add tsc-alias step to make sure all imports are ESM-compliant at build time